+++ HPXMLtoOpenStudio/resources/lighting.rb
@@ -16,7 +16,7 @@ module Lighting
     lighting = hpxml_bldg.lighting
     unit_multiplier = hpxml_bldg.building_construction.number_of_units
     cfa = hpxml_bldg.building_construction.conditioned_floor_area
-    eri_version = hpxml_header.eri_calculation_version
+    eri_version = hpxml_header.eri_calculation_versions[0]
     n_occ = hpxml_bldg.building_occupancy.number_of_residents
 
     ltg_locns = [HPXML::LocationInterior, HPXML::LocationExterior, HPXML::LocationGarage]