+++ Changelog.md
@@ -1,7 +1,8 @@
 ## OpenStudio-HPXML v1.10.0
 
 __New Features__
-- Allows detailed modeling of electric vehicles (batteries and charging/discharging) as an alternative to the simple EV charging plug load.
+- Allows detailed modeling of electric vehicles (batteries and charging/discharging) using `Vehicles` as an alternative to the simple EV charging plug load.
+- Allows requesting timeseries EnergyPlus output meters (e.g., `--hourly "MainsWater:Facility"`), similar to requesting EnergyPlus output variables.
 
 __Bugfixes__
 - Fixes zero occupants specified for one unit in a whole MF building from being treated like zero occupants for every unit.